+++ .github/workflows/ci.yml
@@ -14,6 +14,6 @@ jobs:
           pip install -r requirements.txt
           pip install pytest mypy
       - name: Run mypy
-        run: mypy --strict batch_manager.py openai_batch.py file_processor.py utils.py app.py
+        run: mypy --strict app/batch_manager.py app/openai_batch.py app/file_processor.py app/utils.py streamlit/app.py
       - name: Run tests
         run: pytest
