+++ .gitignore
@@ -25,8 +25,7 @@ wheels/
 
 # Project-specific directories
 logs/
-batch_data/
-batches/
+app/batches/
 
 # Virtual Environment
 venv/
