+++ app/tests/test_utils.py
@@ -1,5 +1,5 @@
 import os
-from utils import sanitize_input, calculate_batch_size, generate_hash
+from app.utils import sanitize_input, calculate_batch_size, generate_hash
 
 
 def test_sanitize_input():
@@ -28,7 +28,7 @@ if 'pandas' not in sys.modules:
 if 'pytz' not in sys.modules:
     sys.modules['pytz'] = types.SimpleNamespace(timezone=lambda tz: tz)
 
-from utils import sanitize_input
+from app.utils import sanitize_input
 
 class TestUtils(unittest.TestCase):
     def test_sanitize_input_preserves_punctuation(self):
