+++ app/wowrunner.py
@@ -1,9 +1,13 @@
 import asyncio
 from typing import List, Optional
 
-from openai_batch import run_batch, poll_batch_until_complete
-from utils import format_batch_summary, format_batch_results, calculate_cost_estimate
-import prompt_store
+from .openai_batch import run_batch, poll_batch_until_complete
+from .utils import (
+    format_batch_summary,
+    format_batch_results,
+    calculate_cost_estimate,
+)
+from . import prompt_store
 from loguru import logger
